[Refactoring] 디렉토리를 Domain으로 구분
