+++ src/main/java/com/web/app/dto/member/MemberSignUpRequsetDTO.java
@@ -10,6 +10,4 @@ public class MemberSignUpRequsetDTO {
 	private final String passwd;
 	private final String email;
 	private final String role;
-	//그러면 MemberDTO를 따로 둔 상태로 Member 도메인을 두는건가,,?
-	//그러면 계층간 데이터 전송을 할 때는 MemberDTO를 쓰고 비즈니스 로직에서는 Member 도메인 객체인건가??
 }
